Make GSL compilation depend on LGPL flag
