+++ src/gpl/README
@@ -1,0 +1,3 @@
+This folder contains code ported or moved from other projects that are under a 
+GPL license.  We actively do not include any part of this code in ARTS when 
+the flag for LGPL is in place
