+++ src/gpl/gsl_gauss_legendre.h
@@ -1,0 +1,39 @@
+#pragma once
+
+/* Copyright (C) 2003-2012 Oliver Lemke  <olemke@core-dump.info>
+
+   This program is free software; you can redistribute it and/or modify it
+   under the terms of the GNU General Public License as published by the
+   Free Software Foundation; either version 2, or (at your option) any
+   later version.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program; if not, write to the Free Software
+   Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307,
+   USA.
+ */
+
+////////////////////////////////////////////////////////////////////////////
+//   File description
+////////////////////////////////////////////////////////////////////////////
+/**
+  \file   legendre.h
+
+  Contains the code to calculate Legendre polynomials.
+
+  \author Oliver Lemke
+  \date 2003-08-14
+  */
+
+#include "arts.h"
+#include "grids.h"
+#include "matpack_data.h"
+
+namespace GSL::Integration {
+bool GaussLegendre(Vector &x, Vector &w, Index n);
+}  // namespace GSL::Integration
