+++ src/gpl/gsl_gauss_legendre_lgpl.cc
@@ -1,0 +1,6 @@
+#include "debug.h"
+#include "gsl_gauss_legendre.h"
+
+bool GSL::Integration::GaussLegendre(Vector&, Vector&, Index) {
+  ARTS_USER_ERROR("This function is not available because ARTS has been compiled in LGPL mode")
+}
