+++ src/legendre.h
@@ -1,50 +1,7 @@
-/* Copyright (C) 2003-2012 Oliver Lemke  <olemke@core-dump.info>
+#pragma once
 
-   This program is free software; you can redistribute it and/or modify it
-   under the terms of the GNU General Public License as published by the
-   Free Software Foundation; either version 2, or (at your option) any
-   later version.
-
-   This program is distributed in the hope that it will be useful,
-   but WITHOUT ANY WARRANTY; without even the implied warranty of
-   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-   GNU General Public License for more details.
-
-   You should have received a copy of the GNU General Public License
-   along with this program; if not, write to the Free Software
-   Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307,
-   USA.
- */
-
-////////////////////////////////////////////////////////////////////////////
-//   File description
-////////////////////////////////////////////////////////////////////////////
-/**
-  \file   legendre.h
-
-  Contains the code to calculate Legendre polynomials.
-
-  \author Oliver Lemke
-  \date 2003-08-14
-  */
-
-#ifndef legendre_h
-#define legendre_h
-
-#include "arts.h"
 #include "grids.h"
 #include "matpack_data.h"
-
-struct gsl_integration_glfixed_table {
-  size_t n;        /* number of points */
-  double* x;       /* Gauss abscissae/points */
-  double* w;       /* Gauss weights for each abscissae */
-  int precomputed; /* high precision abscissae/weights precomputed? */
-};
-
-bool gsl_integration_glfixed_table_alloc(Vector& x, Vector& w, Index n);
-
-
 namespace Legendre {
 
 //! Stores the up (U), south (S), east (E) values of a field relative to the sphere
@@ -54,10 +11,10 @@ struct SphericalField {
   Numeric E{0};
   
   //! Returns the total strength of the field
-  Numeric total() const noexcept {return std::hypot(U, S, E);}
+  [[nodiscard]] Numeric total() const noexcept {return std::hypot(U, S, E);}
   
   //! Returns the total strength of the field
-  Numeric total_horizontal() const noexcept {return std::hypot(S, E);}
+  [[nodiscard]] Numeric total_horizontal() const noexcept {return std::hypot(S, E);}
   
   //! Always construct to zeroes explicitly
   constexpr SphericalField() noexcept = default;
@@ -109,5 +66,3 @@ SphericalField schmidt_fieldcalc(const Matrix& g, const Matrix& h, const Numeric
  */
 MatrixOfSphericalField schmidt_fieldcalc(const Matrix& g, const Matrix& h, const Numeric r0, const Vector& r, const Numeric lat, const Vector& lon) ARTS_NOEXCEPT;
 } // namespace Legendre
-
-#endif /* legendre_h */
